Support negative height in BMP files
The BMP file format standard allows for the bitmap height to be defined as a negative number. When negative, it means that instead of the origin being at the lower left it is instead at the upper left when the height is negative. This file change supports those images where the origin is in the upper left.