Adds support for RLE compression to BMPs
Fixes adafruit/Adafruit_CircuitPython_ImageLoad#20

RLE compression is allowed on 4-bit and 8-bit BMP images. This commit adds
handling for compression levels 1 and 2.

It adds 2 tiny test files for the 4-bit and 8-bit depths, since the
algorithm differs slightly between them. It also includes a version
of the color_wheel.bmp encoded with 8-bit RLE. I cut the image in half
horizontally so that it would load on the PyGamer, which couldn't
allocate the full Bitmap (presumably due to heap fragmentation).